SLEMA-79 Moved class to another file
